Add files via upload
This is a Content Based Product Recommendation System, which uses TfidVectorizer to show the top 10 products which are similar to the product that has been entered by the user. A cosine matrix is generated from the Tfid matrix. The Tfid score assigns the importance of different keywords in the description( excluding stopwords).According to the importance of the words assigned, the dataframe is traversed and the output is displayed.